+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 - Ensure width of animations is divisible by pixel count
 - Update `configure.ac` to use `-lX11` for `raylib` build on `linux`, `macOS`
 - Avoid drawing control hints over notifications
-- Avoid breaking `win_condition` on debug floor regen
+- Avoid breaking `win_condition`, `player` name and class on debug floor regen
 - Use `BLACK` for `rb_notifications` rect
 
 ## [1.4.14] - 2025-12-28